// src/enums/common/RoutePathsEnum.ts
export enum RoutePathsEnum {
  ACCOUNT_ID = ':accountId',

  SYSTEM_WALLETS = 'system-wallets',

  USER_ID = ':userId',

  EMPTY = '',

  ROOT = '/',

  SIGN_IN = 'sign-in',

  MAINTENANCE = 'maintenance',

  LIMITS = 'limits',

  FEE = 'fee',

  PATENTS = 'patents',

  WITHDRAW = 'withdraw',

  ALGORITHM = 'algorithm',

  ALERTS = 'alerts',

  SETTINGS = 'settings',

  STATISTIC = 'statistic',

  USERS = 'users',

  SANDBOX = 'sandbox',

  ORDERS = 'orders',

  REFERRAL = 'referral',

  PAIRS = 'pairs',

  TRADES = 'trades',

  CURRENCIES = 'currencies',

  AI_TRADING = 'ai-trading',

  PROFITS = 'profits',

  HISTORY = 'history',

  TRADING_WALLET = 'trading-wallet',

  BETS = 'bets',

  LOGIN_HISTORY = 'login-history',

  TRANSACTIONS = 'transactions',

  PROFIT_TRANSACTIONS = 'profit-transactions',

  REFERRAL_TRANSACTIONS = 'referral-transactions',

  WITHDRAWAL_REQUESTS = 'withdrawal-requests',

  TURNOVER_TRANSACTIONS = 'turnover-transactions',

  DISTRIBUTION_TRANSACTIONS = 'distribution-transactions',

  WITHDRAW_TRANSACTIONS = 'withdraw-transactions',
}